develop to master

+++ src/Controller/CoordinateController.php
@@ -5,10 +5,12 @@ namespace App\Controller;
 use App\Entity\Coordinate;
 use App\Form\CoordinateType;
 use App\Repository\CoordinateRepository;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
@@ -18,7 +20,8 @@ use Symfony\Component\Routing\Annotation\Route;
 class CoordinateController extends Controller
 {
     /**
-     * @Route("/", name="coordinate_index", methods="GET")
+     * @Route("/", name="coordinate_index")
+     * @Method("GET")
      */
     public function index(CoordinateRepository $coordinateRepository): Response
     {
@@ -26,11 +29,12 @@ class CoordinateController extends Controller
     }
 
     /**
-     * @Route("/new", name="coordinate_new", methods="GET|POST")
+     * @Route("/new", name="coordinate_new")
+     * @Method({"GET", "POST"})
      */
     public function new(Request $request): Response
     {
-        $coordinate = new Coordinate();
+        $coordinate = new Coordinate($this->getUser());
         $form = $this->createForm(CoordinateType::class, $coordinate);
         $form->handleRequest($request);
 
@@ -49,7 +53,8 @@ class CoordinateController extends Controller
     }
 
     /**
-     * @Route("/{id}", name="coordinate_show", methods="GET")
+     * @Route("/{id}", name="coordinate_show")
+     * @Method("GET")
      */
     public function show(Coordinate $coordinate, Request $request): Response
     {
@@ -68,10 +73,15 @@ class CoordinateController extends Controller
     }
 
     /**
-     * @Route("/{id}/edit", name="coordinate_edit", methods="GET|POST")
+     * @Route("/{id}/edit", name="coordinate_edit")
+     * @Method({"GET", "POST"})
      */
     public function edit(Request $request, Coordinate $coordinate): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN', $this->getUser()) && $coordinate->getUser() !== $this->getUser()) {
+            throw new AccessDeniedHttpException();
+        }
+
         $form = $this->createForm(CoordinateType::class, $coordinate);
         $form->handleRequest($request);
 
@@ -88,10 +98,15 @@ class CoordinateController extends Controller
     }
 
     /**
-     * @Route("/{id}", name="coordinate_delete", methods="DELETE")
+     * @Route("/{id}", name="coordinate_delete")
+     * @Method("DELETE")
      */
     public function delete(Request $request, Coordinate $coordinate): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN', $this->getUser()) && $coordinate->getUser() !== $this->getUser()) {
+            throw new AccessDeniedHttpException();
+        }
+
         if ($this->isCsrfTokenValid('delete'.$coordinate->getId(), $request->request->get('_token'))) {
             $em = $this->getDoctrine()->getManager();
             $em->remove($coordinate);
